+++ lib/avx10_t1/mb_mgr_avx10_t1.c
@@ -261,8 +261,8 @@
 #define FLUSH_JOB_ZUC_NEA6  flush_job_zuc_nea6_gfni_avx512
 
 /* ZUC NIA6 */
-#define SUBMIT_JOB_ZUC_NIA6 submit_job_zuc_nia6_gfni_sse
-#define FLUSH_JOB_ZUC_NIA6  flush_job_zuc_nia6_gfni_sse
+#define SUBMIT_JOB_ZUC_NIA6 submit_job_zuc_nia6_gfni_avx512
+#define FLUSH_JOB_ZUC_NIA6  flush_job_zuc_nia6_gfni_avx512
 
 /* ZUC NCA6 */
 #define SUBMIT_JOB_ZUC_NCA6 submit_job_zuc_nca6_gfni_sse
@@ -439,7 +439,7 @@ reset_ooo_mgrs(IMB_MGR *state)
         ooo_mgr_zuc_reset(state->zuc_eea3_ooo, 16);
         ooo_mgr_zuc_reset(state->zuc_eia3_ooo, 16);
         ooo_mgr_zuc_reset(state->zuc_nea6_ooo, 16);
-        ooo_mgr_zuc_reset(state->zuc_nia6_ooo, 4);
+        ooo_mgr_zuc_reset(state->zuc_nia6_ooo, 16);
         ooo_mgr_zuc_reset(state->zuc_nca6_ooo, 4);
 
         /* Init HMAC/SHA1 out-of-order fields */